+++ src/machinelearningservices/azext_mlv2/manual/_params/_model_params.py
@@ -64,7 +64,8 @@ def load_model_params(self):
         c.argument(
             "default_deployment_template",
             options_list=["--default-deployment-template"],
-            help="Default deployment template for the model.",
+            help="Default deployment template for the model. Can be overridden using --set with the format: "
+                 "--set default_deployment_template.asset_id=\"azureml://registries/<registry-name>/deploymenttemplates/<template-name>/versions/<version>\"",
         )
 
     with self.argument_context("ml model show") as c: